requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.0.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
marshmallow==3.20.1
marshmallow-dataclass==8.5.14
mypy==1.4.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
